Styled Nav from Antd, addFriend protected route and pass history prop
