+++ friends/src/components/AddFriend.jsx
@@ -2,6 +2,7 @@ import React, { useState } from "react";
 import { axiosWithAuth } from "../auth/axiosWithAuth";
 
 const AddFriendForm = props => {
+  console.log(props);
   const [friend, setFriend] = useState({ name: "", age: "", email: "" });
   const onInputChange = evt => {
     console.log(evt.target.value);
